start coor

+++ xbee_vari_type.h
@@ -2,11 +2,11 @@
 #define _xbee_vari_type_h_
 
 typedef unsigned char uint8;
-typedef unsigned int  uint16;
+typedef unsigned int  uint32;
 typedef char int8;
-typedef int  int16;
-typedef unsigned long uint32;
-typedef long int32;
+typedef int  int32;
+typedef short int int16;
+typedef unsigned short int uint16;
 typedef unsigned long long uint64;
 typedef long long int64;
